requirements.txt
anthropic>=0.42.0
fastapi>=0.115.0
uvicorn>=0.34.0
chromadb>=0.5.0
pypdf>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
